refactor(fs): extract system opener module
